+++ cloudos/queue/__init__.py
@@ -1,0 +1,8 @@
+"""
+Functions and classes related to job queues.
+"""
+
+from .queue import Queue
+
+
+__all__ = ['queue']